+++ packages/analyzer/CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+* Legacy Polymer function component features will no longer have a `_template`
+  property.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.1.0] - 2018-07-25