+++ packages/gen-typescript-declarations/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+* Legacy Polymer function components will no longer have a `_template`
+  property.
 
 ## [1.4.0] - 2018-07-25
 - Support for ES module imports and exports.